clean up and remove dead code, move different components to their own files, remove common.h
