+++ wolfhsm/wh_cryptocb.h
@@ -20,7 +20,6 @@
  */
 #ifndef WOLFHSM_CRYPTOCB_H
 #define WOLFHSM_CRYPTOCB_H
-#include "wolfhsm/common.h"
 
 #ifdef __cplusplus
     extern "C" {
