+++ wolfhsm/wh_server_she.h
@@ -1,0 +1,9 @@
+#include "wolfhsm/wh_server.h"
+
+#ifndef WOLFHSM_WH_SERVER_SHE_H
+#define WOLFHSM_WH_SERVER_SHE_H
+#if 0
+int hsmHandleSHE(WOLFHSM_CTX* ctx, wh_Packet* packet,
+    NvmMetaData* meta);
+#endif
+#endif
